+++ crates/ruma-state-res/src/state_res/tests.rs
@@ -428,7 +428,7 @@ fn test_reverse_topological_power_sort() {
     };
 
     let res = crate::reverse_topological_power_sort(&graph, |_id| {
-        Ok((int!(0), MilliSecondsSinceUnixEpoch(uint!(0))))
+        Ok((int!(0).into(), MilliSecondsSinceUnixEpoch(uint!(0))))
     })
     .unwrap();