+++ crates/ruma-state-res/src/test_utils.rs
@@ -85,7 +85,7 @@ pub(crate) fn do_check(
     // Resolve the current state and add it to the state_at_event map then continue
     // on in "time"
     for node in crate::reverse_topological_power_sort(&graph, |_id| {
-        Ok((int!(0), MilliSecondsSinceUnixEpoch(uint!(0))))
+        Ok((int!(0).into(), MilliSecondsSinceUnixEpoch(uint!(0))))
     })
     .unwrap()
     {